+++ ackpine-api/api-main/src/main/kotlin/ru/solrudev/ackpine/uninstaller/parameters/UninstallParameters.kt
@@ -16,6 +16,10 @@
 
 package ru.solrudev.ackpine.uninstaller.parameters
 
+import ru.solrudev.ackpine.plugability.AckpinePlugin
+import ru.solrudev.ackpine.plugability.AckpinePluginCache
+import ru.solrudev.ackpine.plugability.AckpinePluginContainer
+import ru.solrudev.ackpine.plugability.AckpinePluginRegistry
 import ru.solrudev.ackpine.session.parameters.Confirmation
 import ru.solrudev.ackpine.session.parameters.ConfirmationAware
 import ru.solrudev.ackpine.session.parameters.NotificationData
@@ -31,6 +35,13 @@ public class UninstallParameters private constructor(
 	public val packageName: String,
 
 	/**
+	 * Type of the package uninstaller implementation.
+	 *
+	 * Default value is [UninstallerType.DEFAULT].
+	 */
+	public val uninstallerType: UninstallerType,
+
+	/**
 	 * A strategy for handling user's confirmation of installation or uninstallation.
 	 *
 	 * Default strategy is [Confirmation.DEFERRED].
@@ -44,7 +55,12 @@ public class UninstallParameters private constructor(
 	 *
 	 * Ignored when [confirmation] is [Confirmation.IMMEDIATE].
 	 */
-	public override val notificationData: NotificationData
+	public override val notificationData: NotificationData,
+
+	/**
+	 * [Plugins][AckpinePlugin] applied to the uninstall session.
+	 */
+	public val pluginContainer: AckpinePluginContainer
 ) : ConfirmationAware {
 
 	override fun equals(other: Any?): Boolean {
@@ -52,33 +68,51 @@ public class UninstallParameters private constructor(
 		if (javaClass != other?.javaClass) return false
 		other as UninstallParameters
 		if (packageName != other.packageName) return false
+		if (uninstallerType != other.uninstallerType) return false
 		if (confirmation != other.confirmation) return false
 		if (notificationData != other.notificationData) return false
+		if (pluginContainer != other.pluginContainer) return false
 		return true
 	}
 
 	override fun hashCode(): Int {
 		var result = packageName.hashCode()
+		result = 31 * result + uninstallerType.hashCode()
 		result = 31 * result + confirmation.hashCode()
 		result = 31 * result + notificationData.hashCode()
+		result = 31 * result + pluginContainer.hashCode()
 		return result
 	}
 
 	override fun toString(): String {
-		return "UninstallParameters(packageName=$packageName, confirmation=$confirmation, " +
-				"notificationData=$notificationData)"
+		return "UninstallParameters(" +
+				"packageName='$packageName', " +
+				"uninstallerType=$uninstallerType, " +
+				"confirmation=$confirmation, " +
+				"notificationData=$notificationData, " +
+				"pluginContainer=$pluginContainer" +
+				")"
 	}
 
 	/**
 	 * Builder for [UninstallParameters].
 	 */
-	public class Builder(packageName: String) : ConfirmationAware {
+	public class Builder(packageName: String) : ConfirmationAware, AckpinePluginRegistry<Builder> {
+
+		private val plugins = mutableMapOf<Class<out AckpinePlugin<*>>, AckpinePlugin.Parameters>()
 
 		/**
 		 * Name of the package to be uninstalled.
 		 */
 		public var packageName: String = packageName
 			private set
+
+		/**
+		 * Type of the package uninstaller implementation.
+		 *
+		 * Default value is [UninstallerType.DEFAULT].
+		 */
+		public var uninstallerType: UninstallerType = UninstallerType.DEFAULT
 
 		/**
 		 * A strategy for handling user's confirmation of installation or uninstallation.
@@ -106,6 +140,13 @@ public class UninstallParameters private constructor(
 		}
 
 		/**
+		 * Sets [UninstallParameters.uninstallerType].
+		 */
+		public fun setUninstallerType(uninstallerType: UninstallerType): Builder = apply {
+			this.uninstallerType = uninstallerType
+		}
+
+		/**
 		 * Sets [UninstallParameters.confirmation].
 		 */
 		public fun setConfirmation(confirmation: Confirmation): Builder = apply {
@@ -119,11 +160,29 @@ public class UninstallParameters private constructor(
 			this.notificationData = notificationData
 		}
 
+		override fun <Params : AckpinePlugin.Parameters> usePlugin(
+			plugin: Class<out AckpinePlugin<Params>>,
+			parameters: Params
+		): Builder = apply {
+			plugins[plugin] = parameters
+		}
+
+		override fun usePlugin(plugin: Class<out AckpinePlugin<AckpinePlugin.Parameters.None>>): Builder = apply {
+			plugins[plugin] = AckpinePlugin.Parameters.None
+		}
+
 		/**
 		 * Constructs a new instance of [UninstallParameters].
 		 */
 		public fun build(): UninstallParameters {
-			return UninstallParameters(packageName, confirmation, notificationData)
+			val pluginContainer = AckpinePluginContainer.from(plugins)
+			val pluginInstances = pluginContainer
+				.getPlugins()
+				.map { (pluginClass, _) -> AckpinePluginCache.get(pluginClass) }
+			for (plugin in pluginInstances) {
+				plugin.apply(this)
+			}
+			return UninstallParameters(packageName, uninstallerType, confirmation, notificationData, pluginContainer)
 		}
 	}
 }