+++ ackpine-core/src/main/kotlin/ru/solrudev/ackpine/impl/services/PackageInstallerService.kt
@@ -14,15 +14,16 @@
  * limitations under the License.
  */
 
-package ru.solrudev.ackpine.impl.installer
+package ru.solrudev.ackpine.impl.services
 
+import android.content.Context
 import android.content.IntentSender
 import android.content.pm.PackageInstaller
 import android.os.Build
 import android.os.Handler
+import android.os.Process
 import androidx.annotation.RequiresApi
 import androidx.annotation.RestrictTo
-import ru.solrudev.ackpine.impl.installer.PackageInstallerService.Session
 import ru.solrudev.ackpine.impl.plugability.AckpineService
 import ru.solrudev.ackpine.plugability.AckpinePlugin
 import java.io.Closeable
@@ -84,6 +85,12 @@ public interface PackageInstallerService : AckpineService {
 	public fun abandonSession(sessionId: Int)
 
 	/**
+	 * @param ackpineSessionId ID of the Ackpine install session.
+	 * @see PackageInstaller.uninstall
+	 */
+	public fun uninstall(packageName: String, statusReceiver: IntentSender, ackpineSessionId: UUID)
+
+	/**
 	 * A facade for [PackageInstaller.Session].
 	 */
 	@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
@@ -124,7 +131,8 @@ public interface PackageInstallerService : AckpineService {
  */
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 @RequiresApi(Build.VERSION_CODES.LOLLIPOP)
-public class PackageInstallerSessionWrapper(private val session: PackageInstaller.Session) : Session {
+public class PackageInstallerSessionWrapper(private val session: PackageInstaller.Session) :
+	PackageInstallerService.Session {
 	override fun openWrite(name: String, offsetBytes: Long, lengthBytes: Long): OutputStream =
 		session.openWrite(name, offsetBytes, lengthBytes)
 
@@ -177,4 +185,14 @@ internal class PackageInstallerWrapper(
 		packageInstaller.unregisterSessionCallback(callback)
 
 	override fun abandonSession(sessionId: Int) = packageInstaller.abandonSession(sessionId)
+
+	override fun uninstall(packageName: String, statusReceiver: IntentSender, ackpineSessionId: UUID) =
+		packageInstaller.uninstall(packageName, statusReceiver)
+
+	internal companion object {
+		@JvmSynthetic
+		internal fun default(context: Context) = lazy {
+			PackageInstallerWrapper(context.packageManager.packageInstaller, Process.myUid())
+		}
+	}
 }