+++ ackpine-ktx/src/main/kotlin/ru/solrudev/ackpine/uninstaller/parameters/UninstallParametersDsl.kt
@@ -16,21 +16,31 @@
 
 package ru.solrudev.ackpine.uninstaller.parameters
 
+import ru.solrudev.ackpine.plugability.AckpinePlugin
+import ru.solrudev.ackpine.plugability.AckpinePluginRegistryDsl
 import ru.solrudev.ackpine.session.parameters.Confirmation
 import ru.solrudev.ackpine.session.parameters.ConfirmationDsl
 import ru.solrudev.ackpine.session.parameters.NotificationData
 import ru.solrudev.ackpine.session.parameters.SessionParametersDsl
+import kotlin.reflect.KClass
 
 /**
  * DSL allowing to configure [parameters for creating uninstall session][UninstallParameters].
  */
 @SessionParametersDsl
-public interface UninstallParametersDsl : ConfirmationDsl {
+public interface UninstallParametersDsl : ConfirmationDsl, AckpinePluginRegistryDsl {
 
 	/**
 	 * Name of the package to be uninstalled.
 	 */
 	public var packageName: String
+
+	/**
+	 * Type of the package uninstaller implementation.
+	 *
+	 * Default value is [UninstallerType.DEFAULT].
+	 */
+	public var uninstallerType: UninstallerType
 }
 
 @PublishedApi
@@ -55,6 +65,23 @@ internal class UninstallParametersDslBuilder(packageName: String) : UninstallPar
 		set(value) {
 			builder.setPackageName(value)
 		}
+
+	override var uninstallerType: UninstallerType
+		get() = builder.uninstallerType
+		set(value) {
+			builder.setUninstallerType(value)
+		}
+
+	override fun <Params : AckpinePlugin.Parameters> usePlugin(
+		plugin: KClass<out AckpinePlugin<Params>>,
+		parameters: Params
+	) {
+		builder.usePlugin(plugin.java, parameters)
+	}
+
+	override fun usePlugin(plugin: KClass<out AckpinePlugin<AckpinePlugin.Parameters.None>>) {
+		builder.usePlugin(plugin.java)
+	}
 
 	fun build() = builder.build()
 }