+++ build-logic/src/main/kotlin/ru/solrudev/ackpine/gradle/app/AckpineKotlinSamplePlugin.kt
@@ -20,7 +20,9 @@ import com.android.build.api.dsl.ApplicationExtension
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.apply
+import org.gradle.kotlin.dsl.assign
 import org.gradle.kotlin.dsl.configure
+import org.jetbrains.kotlin.gradle.dsl.JvmDefaultMode
 import org.jetbrains.kotlin.gradle.dsl.KotlinAndroidExtension
 
 public class AckpineKotlinSamplePlugin : Plugin<Project> {
@@ -39,7 +41,8 @@ public class AckpineKotlinSamplePlugin : Plugin<Project> {
 
 	private fun Project.configureKotlin() = extensions.configure<KotlinAndroidExtension> {
 		compilerOptions {
-			freeCompilerArgs.addAll("-Xjvm-default=all", "-Xconsistent-data-class-copy-visibility")
+			jvmDefault = JvmDefaultMode.NO_COMPATIBILITY
+			freeCompilerArgs.add("-Xconsistent-data-class-copy-visibility")
 		}
 	}
 }