+++ sample-ktx/src/main/kotlin/ru/solrudev/ackpine/sample/uninstall/UninstallUiState.kt
@@ -18,5 +18,6 @@ package ru.solrudev.ackpine.sample.uninstall
 
 data class UninstallUiState(
 	val isLoading: Boolean = false,
-	val applications: List<ApplicationData> = emptyList()
+	val applications: List<ApplicationData> = emptyList(),
+	val failure: String? = null
 )